- Finished implementing new LevelManager which can now dynamically calculate the level for each karma score from a dynamically generated LevelLookup. It also contains a LevelThresholdWeighting property which controls how easy/hard it is to level up.
- As part of the above work, Karma has now been changed from an int to a long in AvatarDetail, IAvatarDetail interface & KarmaAkashicRecord.

- Added new AvatarBasicView which will be used in new api calls in the avatar api to list avatars to standard users (and will be used for the OASIS Avatar Directory, avatar lookup/searches, leader boards, etc.

- Fixed bugs in NextGenSoftware.OASIS.API.Core.TestHarness.

- Added new LevelManager tests to NextGenSoftware.OASIS.API.Core.TestHarness.

- Fixed ref issue for the OASIS Logo in NextGenSoftware.OASIS.API.Native.Integrated.EndPoint.

- Made all methods and properties static in OASISAPI helper in NextGenSoftware.OASIS.API.Native.Integrated.EndPoint.

- Made Karma long in SolanaAvatarDetailDto in SolanaOASIS Provider.

- Made karma long in AvatarDetailModel and KarmaAkashicRecordModel in SQLLiteDBOASIS Provider.

- Made karma long in AvatarDetail in MongoDBOASIS Provider.

+++ NextGenSoftware.OASIS.API.Core/Interfaces/Avatar/IAvatarDetail.cs
@@ -48,7 +48,7 @@ namespace NextGenSoftware.OASIS.API.Core.Interfaces
         HumanDesign HumanDesign { get; set; }
         string Portrait { get; set; }
         List<InventoryItem> Inventory { get; set; }
-        int Karma { get; set; }
+        long Karma { get; set; }
         List<KarmaAkashicRecord> KarmaAkashicRecords { get; set; }
         string Landline { get; set; }
         int Level { get; }
