+++ NextGenSoftware.OASIS.API.Providers.SOLANAOASIS/Entities/Models/SolanaAvatarDetailDto.cs
@@ -4,7 +4,7 @@
     {
         public string Address { get; set; }
         public string Mobile { get; set; }
-        public int Karma { get; set; }
+        public long Karma { get; set; }
         public int Xp { get; set; }
     }
 }
